refactor[palantir]: restructure package
